refactor mock-dsl-jvm to re-use ValueClassSupport code, and add TODOs to de-dupe them #857
